Removed `client` module
